chore: move proc macro implementation to separate modules
